requirements.txt
torch
torchvision
torchaudio
scikit-learn
matplotlib
seaborn
pandas
tqdm
streamlit
opencv-python
pillow
grad-cam
